Move code into a more common module to reduce cross module deps.
