+++ tests/test_git.py
@@ -2,27 +2,18 @@ from __future__ import absolute_import
 import lintreview.git as git
 import os
 from .test_github import config
+from . import (
+    setup_repo,
+    teardown_repo,
+    clone_path,
+    cant_write_to_test
+)
 from nose.tools import eq_, raises, assert_in, with_setup
 from unittest import skipIf
 
 settings = {
-    'WORKSPACE': './tests'
+    'WORKSPACE': clone_path + '/tests'
 }
-clone_path = settings['WORKSPACE'] + '/test_clone'
-cant_write_to_test = not(os.access(os.path.abspath('./tests'), os.W_OK))
-
-
-def setup_repo():
-    git.clone_or_update(
-        config,
-        'git://github.com/markstory/lint-review.git',
-        clone_path,
-        'master')
-
-
-def teardown_repo():
-    if git.exists(clone_path):
-        git.destroy(clone_path)
 
 
 def noop():
@@ -126,7 +117,8 @@ def test_diff__files_list():
 @skipIf(cant_write_to_test, 'Cannot write to ./tests skipping')
 @raises(IOError)
 def test_diff__non_git_path():
-    git.diff(settings['WORKSPACE'] + '/../../')
+    path = os.path.abspath(clone_path + '/../../../')
+    git.diff(path)
 
 
 @skipIf(cant_write_to_test, 'Cannot write to ./tests skipping')
@@ -163,7 +155,8 @@ def test_apply_cached__bad_patch():
 @skipIf(cant_write_to_test, 'Cannot write to ./tests skipping')
 @raises(IOError)
 def test_apply_cached__non_git_path():
-    git.apply_cached(settings['WORKSPACE'] + '/../../', 'not a patch')
+    path = os.path.abspath(clone_path + '/../../')
+    git.apply_cached(path, 'not a patch')
 
 
 @skipIf(cant_write_to_test, 'Cannot write to ./tests skipping')
